requirements.txt
python-docx==0.8.11
pandas==2.0.3
google-generativeai==0.3.1
python-dotenv==1.0.0
